Move the two resource classes to separate package
This further helps protect the package-protected fields in Resources.

< organisation, interface
