+++ src/main/java/nl/tudelft/model/Level.java
@@ -7,8 +7,8 @@ import nl.tudelft.model.pickups.Pickup;
 import nl.tudelft.model.pickups.weapon.Projectile;
 import nl.tudelft.semgroup4.Modifiable;
 import nl.tudelft.semgroup4.Renderable;
-import nl.tudelft.semgroup4.Resources;
 import nl.tudelft.semgroup4.Updateable;
+import nl.tudelft.semgroup4.Resources.Resources;
 import nl.tudelft.semgroup4.util.Helpers;
 
 import org.newdawn.slick.GameContainer;
