+++ src/main/java/nl/tudelft/model/bubble/Bubble1.java
@@ -1,6 +1,6 @@
 package nl.tudelft.model.bubble;
 
-import nl.tudelft.semgroup4.ResourcesWrapper;
+import nl.tudelft.semgroup4.Resources.ResourcesWrapper;
 
 /**
  * One of the Bubble types. This is a regular bubble of size 1.
