+++ src/main/java/nl/tudelft/model/pickups/powerup/ShieldPowerup.java
@@ -2,7 +2,7 @@ package nl.tudelft.model.pickups.powerup;
 
 import nl.tudelft.model.Player;
 import nl.tudelft.semgroup4.Modifiable;
-import nl.tudelft.semgroup4.Resources;
+import nl.tudelft.semgroup4.Resources.Resources;
 
 import org.newdawn.slick.Color;
 import org.newdawn.slick.GameContainer;
