+++ src/main/java/nl/tudelft/model/pickups/powerup/SpeedPowerup.java
@@ -2,7 +2,7 @@ package nl.tudelft.model.pickups.powerup;
 
 import nl.tudelft.model.Player;
 import nl.tudelft.semgroup4.Modifiable;
-import nl.tudelft.semgroup4.Resources;
+import nl.tudelft.semgroup4.Resources.Resources;
 
 import org.newdawn.slick.SlickException;
