+++ src/main/java/nl/tudelft/semgroup4/MainApp.java
@@ -2,6 +2,8 @@ package nl.tudelft.semgroup4;
 
 import java.io.File;
 
+import nl.tudelft.semgroup4.Resources.Resources;
+
 import org.lwjgl.LWJGLUtil;
 import org.newdawn.slick.AppGameContainer;
 import org.newdawn.slick.GameContainer;
