+++ src/test/java/nl/tudelft/model/GameObjectTest.java
@@ -1,8 +1,8 @@
 package nl.tudelft.model;
 
 import static org.junit.Assert.assertEquals;
+import nl.tudelft.semgroup4.Resources.Resources;
 
-import nl.tudelft.semgroup4.Resources;
 import org.junit.Test;
 import org.newdawn.slick.SlickException;
