+++ src/test/java/nl/tudelft/model/GameTest.java
@@ -11,7 +11,7 @@ import static org.mockito.Mockito.when;
 import java.util.LinkedList;
 
 import nl.tudelft.model.pickups.weapon.Projectile;
-import nl.tudelft.semgroup4.Resources;
+import nl.tudelft.semgroup4.Resources.Resources;
 import nl.tudelft.semgroup4.collision.CollisionHandler;
 
 import org.junit.Test;
