+++ src/test/java/nl/tudelft/model/pickups/weapon/WeaponTest.java
@@ -9,11 +9,10 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
-
 import nl.tudelft.model.AbstractOpenGLTestCase;
 import nl.tudelft.model.Game;
 import nl.tudelft.model.Player;
-import nl.tudelft.semgroup4.Resources;
+import nl.tudelft.semgroup4.Resources.Resources;
 
 import org.junit.Test;
 import org.newdawn.slick.Input;
