+++ src/test/java/nl/tudelft/semgroup4/util/AudioTest.java
@@ -2,8 +2,7 @@ package nl.tudelft.semgroup4.util;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-
-import nl.tudelft.semgroup4.Resources;
+import nl.tudelft.semgroup4.Resources.Resources;
 import nl.tudelft.semgroup4.util.Audio;
 
 import org.junit.Test;
